CHANGED FUNCTIONS IN ORDER TO TREAT BASELINE CASE.
Created a parameter 'referenceCase' in Generalparameters in order to flag that the current run is a baseline/reference case.

In this case, demand will be estimated using TMY data. Hydro potentials will be the historic and static ones. Curtailments will not be simulated.

In order to avoid errors in some simulations, we still need to inform a valid GCM name at the list of GCMS at curtailment parameters. This needs to be improved...

Still need to improve this logic...