WRITE UC RESULTS EVERY 10 DAYS. CHANGED STRUCTURE OF GENERALPARAMETERS CLASS.
RIPSMasterScript17Nov2017.py: Now the loop with daily UC simulations will dump
accumulated results every 10 day.

Generalparameters.py: changed function load of this class. Now it reads the the
txt file and uses setattr to set values of attributes. Now we have to be careful
that the names of the attributes in the txt file are the same as the names in
the definition of the class